easy/flood_fill_test.go
package easy

import (
	"reflect"
	"testing"
)

func Test_floodFill(t *testing.T) {
	type args struct {
		image    [][]int
		sr       int
		sc       int
		newColor int
	}
	tests := []struct {
		name string
		args args
		want [][]int
	}{
		{
			name:"one",
			args:args{
				image:    [][]int{[]int{0, 0, 0}, []int{0, 1, 1}},
				sr:       1,
				sc:       1,
				newColor: 1,
			},
			want: [][]int{[]int{0, 0, 0}, []int{0, 1, 1}},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := floodFill(tt.args.image, tt.args.sr, tt.args.sc, tt.args.newColor); !reflect.DeepEqual(got, tt.want) {
				t.Errorf("floodFill() = %v, want %v", got, tt.want)
			}
		})
	}
}